+++ src/moduler/aktivitet/etikett/SokeStatusEtikett.tsx
